backend/requirements.txt
django==5.2.12
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-jazzmin==3.0.3
whitenoise==6.9.0
gunicorn==23.0.0
psycopg[binary]==3.2.10
pillow==11.2.1
pywebpush==1.14.1
python-dotenv==1.0.1
